backend/requirements.txt
fastapi
uvicorn
python-dotenv
yt-dlp
langchain
langchain-google-genai
langchain-community
chromadb
pydantic
functions-framework
